+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## latest
 - accept local identifier for user image and return it in JSON
 - accept local identifier for group image and return it in JSON
-
+- FIX bug where slot can not be updated with a location
 
 ## v0.4.2 on t2s2-silvio-dev
 - return only upcoming friends slots sorted by startdate